# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Yeshshree Operations API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/schedules": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Create Schedule",
        "description": "Manual JSON entry path (provisional until the Bajaj sample \u2014 Domain_QA Q1).",
        "operationId": "create_schedule_api_v1_schedules_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "List Schedules",
        "operationId": "list_schedules_api_v1_schedules_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period"
            }
          },
          {
            "name": "customer_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleRead"
                  },
                  "title": "Response List Schedules Api V1 Schedules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedules/upload": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Upload Schedule",
        "description": "Provisional-CSV upload \u2192 parser \u2192 draft schedule. Row errors are returned,\nnever raised \u2014 except when NO row parses, which is a 422.",
        "operationId": "upload_schedule_api_v1_schedules_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_schedule_api_v1_schedules_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleUploadResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedules/{schedule_id}/lines": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Get Schedule Lines",
        "operationId": "get_schedule_lines_api_v1_schedules__schedule_id__lines_get",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleLineRead"
                  },
                  "title": "Response Get Schedule Lines Api V1 Schedules  Schedule Id  Lines Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedules/{schedule_id}/diff": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Get Diff",
        "operationId": "get_diff_api_v1_schedules__schedule_id__diff_get",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedules/{schedule_id}/sanity": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Get Sanity",
        "description": "Recomputed on read \u2014 the planner sees the verdict for the CURRENT config,\nnot a stale snapshot (release stores its own snapshot in schedules.sanity).",
        "operationId": "get_sanity_api_v1_schedules__schedule_id__sanity_get",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanityRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedules/{schedule_id}/release": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Release Schedule",
        "operationId": "release_schedule_api_v1_schedules__schedule_id__release_post",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Schedule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedules/{schedule_id}/re-release": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Re Release Schedule",
        "description": "\u00a711.13 rollback: clone old version as a new one and release it. Never deletes.",
        "operationId": "re_release_schedule_api_v1_schedules__schedule_id__re_release_post",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/line-plans": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Line Plans",
        "operationId": "line_plans_api_v1_plans_line_plans_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "line_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Line Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanRow"
                  },
                  "title": "Response Line Plans Api V1 Plans Line Plans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/ppc-view": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Ppc View",
        "operationId": "ppc_view_api_v1_plans_ppc_view_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PpcLine"
                  },
                  "title": "Response Ppc View Api V1 Plans Ppc View Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/supervisor-view": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Supervisor View",
        "operationId": "supervisor_view_api_v1_plans_supervisor_view_get",
        "parameters": [
          {
            "name": "line_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Line Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupervisorRow"
                  },
                  "title": "Response Supervisor View Api V1 Plans Supervisor View Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_schedule_api_v1_schedules_upload_post": {
        "properties": {
          "customer_id": {
            "type": "integer",
            "title": "Customer Id"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "customer_id",
          "period",
          "file"
        ],
        "title": "Body_upload_schedule_api_v1_schedules_upload_post"
      },
...
      "DiffRead": {
        "properties": {
          "schedule_id": {
            "type": "integer",
            "title": "Schedule Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "diff": {
            "additionalProperties": true,
            "type": "object",
            "title": "Diff"
          }
        },
        "type": "object",
        "required": [
          "schedule_id",
          "version",
          "diff"
        ],
        "title": "DiffRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanRow": {
        "properties": {
          "plan_id": {
            "type": "integer",
            "title": "Plan Id"
          },
          "plan_date": {
            "type": "string",
            "format": "date",
            "title": "Plan Date"
          },
          "line_id": {
            "type": "integer",
            "title": "Line Id"
          },
          "line_name": {
            "type": "string",
            "title": "Line Name"
          },
          "material_id": {
            "type": "integer",
            "title": "Material Id"
          },
          "sap_code": {
            "type": "string",
            "title": "Sap Code"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "planned_qty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Planned Qty"
          },
          "confirmed_good": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confirmed Good"
          },
          "confirmed_reject": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confirmed Reject"
          },
          "remaining": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Remaining"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "plan_date",
          "line_id",
          "line_name",
          "material_id",
          "sap_code",
          "description",
          "revision",
          "planned_qty",
          "confirmed_good",
          "confirmed_reject",
          "remaining"
        ],
        "title": "PlanRow",
        "description": "Frozen response shape: confirmed_* are placeholder zeros until the\nproduction packet joins confirmations; remaining = planned \u2212 confirmed_good."
      },
      "PpcLine": {
        "properties": {
          "line_id": {
            "type": "integer",
            "title": "Line Id"
          },
          "line_name": {
            "type": "string",
            "title": "Line Name"
          },
          "plans": {
            "items": {
              "$ref": "#/components/schemas/PlanRow"
            },
            "type": "array",
            "title": "Plans"
          }
        },
        "type": "object",
        "required": [
          "line_id",
          "line_name",
          "plans"
        ],
        "title": "PpcLine"
      },
...
      "ReleaseReport": {
        "properties": {
          "schedule_id": {
            "type": "integer",
            "title": "Schedule Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "plans_created": {
            "type": "integer",
            "title": "Plans Created"
          },
          "calloffs_created": {
            "type": "integer",
            "title": "Calloffs Created"
          },
          "unmapped": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Unmapped"
          },
          "no_vendor": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "No Vendor"
          },
          "cloned_from_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloned From Version"
          }
        },
        "type": "object",
        "required": [
          "schedule_id",
          "version",
          "plans_created",
          "calloffs_created",
          "unmapped",
          "no_vendor"
        ],
        "title": "ReleaseReport"
      },
      "ReleaseRequest": {
        "properties": {
          "confirm_warnings": {
            "type": "boolean",
            "title": "Confirm Warnings",
            "default": false
          }
        },
        "type": "object",
        "title": "ReleaseRequest"
      },
...
      "SanityRead": {
        "properties": {
          "schedule_id": {
            "type": "integer",
            "title": "Schedule Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "sanity": {
            "additionalProperties": true,
            "type": "object",
            "title": "Sanity"
          }
        },
        "type": "object",
        "required": [
          "schedule_id",
          "version",
          "sanity"
        ],
        "title": "SanityRead"
      },
...
      "ScheduleCreate": {
        "properties": {
          "customer_id": {
            "type": "integer",
            "title": "Customer Id"
          },
          "period": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}$",
            "title": "Period"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/ScheduleLineIn"
            },
            "type": "array",
            "title": "Lines"
          },
          "source_file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source File Id"
          }
        },
        "type": "object",
        "required": [
          "customer_id",
          "period",
          "lines"
        ],
        "title": "ScheduleCreate"
      },
...
      "ScheduleLineRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "schedule_id": {
            "type": "integer",
            "title": "Schedule Id"
          },
          "model_family": {
            "type": "string",
            "title": "Model Family"
          },
          "material_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Id"
          },
          "bucket_date": {
            "type": "string",
            "format": "date",
            "title": "Bucket Date"
          },
          "qty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Qty"
          }
        },
        "type": "object",
        "required": [
          "id",
          "schedule_id",
          "model_family",
          "material_id",
          "bucket_date",
          "qty"
        ],
        "title": "ScheduleLineRead"
      },
      "ScheduleRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "customer_id": {
            "type": "integer",
            "title": "Customer Id"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source_file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source File Id"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "released_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Released By"
          },
          "released_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Released At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_id",
          "period",
          "version",
          "status",
          "source_file_id",
          "created_by",
          "released_by",
          "released_at"
        ],
        "title": "ScheduleRead"
      },
      "ScheduleUploadResult": {
        "properties": {
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "rows_ok": {
            "type": "integer",
            "title": "Rows Ok"
          },
          "row_errors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Row Errors"
          }
        },
        "type": "object",
        "required": [
          "schedule",
          "rows_ok",
          "row_errors"
        ],
        "title": "ScheduleUploadResult",
        "description": "Upload outcome: the created draft (when any row parsed) + EVERY row error."
      },
...
      "SupervisorRow": {
        "properties": {
          "plan_id": {
            "type": "integer",
            "title": "Plan Id"
          },
          "plan_date": {
            "type": "string",
            "format": "date",
            "title": "Plan Date"
          },
          "line_id": {
            "type": "integer",
            "title": "Line Id"
          },
          "line_name": {
            "type": "string",
            "title": "Line Name"
          },
          "material_id": {
            "type": "integer",
            "title": "Material Id"
          },
          "sap_code": {
            "type": "string",
            "title": "Sap Code"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "planned_qty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Planned Qty"
          },
          "confirmed_good": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confirmed Good"
          },
          "confirmed_reject": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confirmed Reject"
          },
          "remaining": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Remaining"
          },
          "production_order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Production Order Id"
          },
          "sap_order_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sap Order No"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "plan_date",
          "line_id",
          "line_name",
          "material_id",
          "sap_code",
          "description",
          "revision",
          "planned_qty",
          "confirmed_good",
          "confirmed_reject",
          "remaining",
          "production_order_id",
          "sap_order_no"
        ],
        "title": "SupervisorRow"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}